Refactor user mock from all test modules into own single module.
