+++ back/__tests__/changeOwnPassword.test.js
@@ -10,47 +10,15 @@ const bcrypt = require('bcryptjs');
 
 const endpoint = '/api/changeownpassword';
 
-// Mock user model
-jest.mock('../models/user', () => {
-  const users = [
-    {name: 'admin',
-      role: 'superuser',
-      id: '123',
-      email: 'usr@email.com',
-      digest: 't',
-      reset_token: null,
-      reset_token_expire: null
-    }
-  ]; 
-
-  const model = {
-    read: async (params) => users.filter((user) => {
-      const searchKeys = Object.keys(params);
-      let match = true;
-      searchKeys.forEach(key => {
-        if (params[key] !== user[key]) {
-          match = false;
-        }
-      });
-      return match;
-    }),
-    create: async (user) => { 
-      users.push(user);
-      return [user];
-    },
-    update: async (current, update) => {
-      const i = users.findIndex(user => user.id == current.id);
-      users[i] = {...users[i], ...update};
-    },
-    clear: () => users.length = 0
-  };
-  
-  return model;
-});
-
+jest.mock('../models/user');
 const userModel = require('../models/user');
 
 describe(`${endpoint}`, () => {
+  beforeEach(() => {
+    jest.resetAllMocks();
+    userModel.clear();
+  });
+
   describe('PUT', () => {
     it(`When a valid token is provided:
     returns code 204 and updates password.`, async () => {
@@ -97,6 +65,25 @@ describe(`${endpoint}`, () => {
 
     it(`When a valid token is provided, user is superuser and updating another users password:
     returns code 204 and updates password.`, async () => {
+      await userModel.create({name: 'admin',
+        role: 'superuser',
+        id: '123',
+        email: 'usr@email.com',
+        digest: 't',
+        reset_token: null,
+        reset_token_expire: null
+      });
+
+      const user = {name: 'normal',
+        role: 'supervisor',
+        id: '223',
+        email: 'usr2@email.com',
+        reset_token: null,
+        reset_token_expire: null
+      };
+
+      await userModel.create({...user, digest: 't'});
+
       const newPassword = 'wordpass4321';
       const res = await request.put(`${endpoint}/223`)
         .set('Cookie', [`token=${jwt.sign({id: '123'}, config.jwt.secret)}`])
